Split RunningMetrics.jsx
This was split into ViewRunningMetrics which is just the graphs that pull from the database, and InputRunningMetrics which contains all text fields and the logic for uploading to the database using Hono. Both are currently living in RunningMetrics.jsx until I move InputRunningMetrics into its own page (along with all over metric inputs)
